requirements.txt
django
psycopg2-binary
gunicorn
flake8
djangorestframework
django-oauth-toolkit
django-filter